Added mappers to mapp the dtos and entity classes or the model classes made changes to the model class removing some fields and adding some updated some repository since datatype had been changed for the id fields

+++ src/main/java/com/peanutbutter/peanutbutter/dtos/BatchProductResponseDto.java
@@ -7,7 +7,7 @@ public class BatchProductResponseDto {
 
     private Long batchProductID;
     private Long batchid;
-    private int productid;
+    private Long productid;
     private String productName;
     private int productQuantity;
     private int productRemQuantity;
@@ -31,11 +31,11 @@ public class BatchProductResponseDto {
         this.batchid = id;
     }
 
-    public int getProductID(){
+    public Long getProductID(){
         return productid;
     }
 
-    public void setProductID(int id){
+    public void setProductID(Long id){
         this.productid = id;
     }
