+++ src/main/java/com/peanutbutter/peanutbutter/repository/ExpenseRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.peanutbutter.peanutbutter.model.Expense;
 
-public interface ExpenseRepository extends JpaRepository<Expense,Integer> {
+public interface ExpenseRepository extends JpaRepository<Expense,Long> {
 
 }
